+++ docs/Whats-New.md
@@ -421,6 +421,8 @@ New:
 - Enhanced Straight trajectory (by CrimRecya)
 - New Disperse trajectory (by CrimRecya)
 - New Engrave trajectory (by CrimRecya)
+- New Parabola trajectory (by CrimRecya)
+- New Tracing trajectory (by CrimRecya)
 - Display factories, superweapons, iron curtain and temporal progress on specific technos (by CrimRecya)
 - Extra warhead multiplier for different houses (by CrimRecya)
 - Units Rotate Turret When in Idle Action (by CrimRecya)